requirements.txt
# This file is used to build the data-pipeline-tools debian package.  Because
# of this, it installs the tools extras dependencies.  All versions should be
# pinned here, because  these packages will be included in a virtualenv
# deployed inside a debian package - so the build must be repeatable.
# This list was originally constructed by running
# `pip install -i https://pypi.yelpcorp.com/simple/ ".[tools]"`
# and `pip freeze > requirements.txt` in a virtualenv.
#
# Specifically, run:
#
# virtualenv tools
# source tools/bin/activate
# pip install --upgrade pip
# pip install -b /nail/tmp -i https://pypi.yelpcorp.com/simple/ ".[tools]"
# pip freeze > requirements.tmp
# deactivate
# rm -rf tools
#
# Note that the data_pipeline package should not appear here, and that cffi
# should be at least 1.2.1 for trusty compatibility.

argparse==1.4.0
boto==2.38.0
bravado==8.3.0
bravado-core==4.5.0
bravado-decorators==0.6.0
cached-property==1.2.0
cffi==1.6.0
characteristic==14.1.0
contextdecorator==0.10.0
crochet==1.4.0
cryptography==1.3.4
dateglob==0.1
enum34==1.0.4
fido==4.0.1
functools32==3.2.3-2
future==0.14.3
futures==3.0.3
geogrid==1.0.9
ipaddress==1.0.14
iso8601==0.1.10
jsonschema==2.5.1
kafka-python==0.9.5.post6
kafka-utils==0.4.2
kazoo==2.2
mysqlclient==1.3.6
ply==3.4
psutil==4.2.0
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==0.14.0
pysensu_yelp==0.2.3
PyStaticConfiguration==0.9.0
pysubnettree==0.23
python-dateutil==2.4.2
pytz==2014.10
PyYAML==3.11
replication-delay-client==1.1.1
repoze.lru==0.6
requests==2.8.1
retrying==1.3.3
send-nsca==0.1.4.1
setproctitle==1.1.8
simplejson==3.6.5
six==1.9.0
SQLAlchemy==0.9.8
subprocess32==3.2.6
swagger-spec-validator==2.0.2
thriftpy==0.1.15
tornado==2.4.1
Twisted==15.4.0
data-pipeline-avro-util==0.2.1
yelp-avro==1.9.2
yelp-batch==0.19.8
yelp-bytes==0.2.0
yelp-cgeom==1.3.1
yelp-clog==2.5.2
yelp-conn==7.1.3
yelp-encodings==0.1.3
yelp-kafka==5.1.1
yelp-lib==11.0.2
yelp-logging==1.4.1
yelp-meteorite==0.2.4
yelp-servlib==4.7.0
yelp-uri==1.1.0
zope.interface==4.1.2
